I created the skeleton of the tags that will be attached to each helper profile. Currently the tags are just colors, but will be changed later as the app becomes more developed. Created a HelperTags class in the ParseServer, and the user in the HelperTags points to the user in the User class in the ParseServer. When you sign up and choose to be a HELPER, the app will take you to a page with a list of checkboxes, and the user can choose to select any colors they want (can select as many as they want). This will then be passed into the server and saved under HelperTags, where the user and the color selected are both saved.
Next step is to start working on the helper profile, where the tags the helper has will appear under the profile. In the future to make it more user friendly, we can have the option to edit tags anytime, but that can be saved for a bit later. Also I don't know if my code is the most clean because I'm essentially hardcoding in all the checkbox colors, but I did create a ParseServer class with all of the colors on there...maybe there is a way I can use this--will think more about that.

Also saw Javier's comments on my last commit, will definitely work on that too. :)